% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_resistance_info.R
\name{add_resistance_info}
\alias{add_resistance_info}
\title{Reistance Genotyping}
\usage{
add_resistance_info(f.dat, resistance_table, all_muts = F)
}
\arguments{
\item{f.dat}{intermediate-annotated data.frame}

\item{resistance_table}{the current version of the resistance db in csv format}
}
\value{
data.frame of resistance variants
}
\description{
Calls resistance for variants provided
}
\keyword{internal}
