% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_variants.R
\name{annotate_variants}
\alias{annotate_variants}
\title{Annotates variants}
\usage{
annotate_variants(f.dat, global)
}
\arguments{
\item{f.dat}{intermediate data.frame}

\item{global}{Package object for consistent runtime variables}
}
\value{
intermediate data.frame with genome level annotation
}
\description{
Adds genome annotation to the intermediate resistance data.frame
}
\keyword{internal}
