% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_resistance.R
\name{call_resistance}
\alias{call_resistance}
\title{CMV Resistance Genotyping from command line}
\usage{
call_resistance(infile = system.file("testdata", "A10.vcf", package =
  "cmvdrg"), all_mutations = FALSE, outdir = "")
}
\arguments{
\item{infile}{the input fasta, vcf or varscan.tab file}
}
\value{
A data.frame containing resistance information for variants identified
}
\description{
Command line version of the website application
Takes as input a VCF, varscan tab or fasta file.
The program assumes variant files are generated realative to Merlin strain.
Fasta files if not Whole Genome, or not aligned / assembled relative to Merlin 
are Processed using MAFFT & snp-sites.
In this case the output files are returned to your working directory.
}
