% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmvdrg_data.R
\name{cmvdrg_data}
\alias{cmvdrg_data}
\title{Returns a dataframe of the cmvdrg database}
\usage{
cmvdrg_data()
}
\value{
dataframe of the cmvdrg database
}
\description{
Returns a dataframe of the cmvdrg database
}
