% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_clin_table.R
\name{make_clin_table}
\alias{make_clin_table}
\title{Clinical information table}
\usage{
make_clin_table(f.dat)
}
\arguments{
\item{f.dat}{the complete resistance data.frame}
}
\value{
a data.table
}
\description{
Produces a data.table with a clinical overview for resistance phenotype against all drugs in that sample
}
