% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_varscan_data.R
\name{read_varscan_data}
\alias{read_varscan_data}
\title{Handles varscan data}
\usage{
read_varscan_data(f.df)
}
\arguments{
\item{f.df}{data.frame of read varscan file}
}
\value{
formatted data.frame
}
\description{
reads varscan files & returns a formatted data.frame
}
\keyword{internal}
