% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_resistance_info.R
\name{add_resistance_info}
\alias{add_resistance_info}
\title{Resistance Genotyping}
\usage{
add_resistance_info(f.dat, resistance_table, all_muts = FALSE,
  anecdotal = FALSE)
}
\arguments{
\item{f.dat}{intermediate-annotated data.frame}

\item{resistance_table}{the current version of the resistance db in csv format}

\item{all_muts}{when TRUE all variants passed are returned even if they conferred no resistance}

\item{anecdotal}{include anectodat database entries in returned results?}
}
\value{
data.frame of resistance variants
}
\description{
Calls resistance for variants provided, joins the variant and resistance tables by mutational change columns. <GENE_A123B>
}
\keyword{internal}
