% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lollipop.R
\name{plot_lollipop}
\alias{plot_lollipop}
\title{Make gene lollipop plot}
\usage{
plot_lollipop(
  f.dat,
  f.gene = "UL54",
  resistance_table = system.file("db", "cmvdrg-db1.csv", package = "cmvdrg")
)
}
\arguments{
\item{f.dat}{resistance data frame from cmvdrg, where all_muts == F}

\item{f.gene}{Which gene to plot}

\item{resistance_table}{location of the csv resistance database}
}
\value{
intermediate data.frame with genome level annotation
}
\description{
Produces figures for the web application.
Explores spatial location of mutations in resistance genes.
}
