% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_input.R
\name{read_input}
\alias{read_input}
\title{Handles input file}
\usage{
read_input(infile, global)
}
\arguments{
\item{infile}{Path to the input file}

\item{global}{Package object for consistent runtime variables}
}
\value{
An intermediate data.frame
}
\description{
This function handles a vcf, or varscan tab variant file, or pases fasta files
onto handle_fasta().
Returns an intermediate data.frame, which contains variant information to be
annotated by annotate_variants().
}
\keyword{internal}
